% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_mode}
\alias{calculate_mode}
\title{calculate mode}
\usage{
calculate_mode(values)
}
\arguments{
\item{values}{the values you want the first primary mode of}
}
\value{
numeric
}
\description{
Calculates the mode of a set of values
}
