% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_feature_class}
\alias{check_feature_class}
\title{check class}
\usage{
check_feature_class(data, feature_column = "feature",
  class_column = "class")
}
\arguments{
\item{data}{the data.frame to work on}

\item{class_column}{which columns has the class}

\item{feature_colum}{which column has the feature}
}
\description{
runs a check that the class can actually be found in the feature
}
