% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{create_correlation_plot}
\alias{create_correlation_plot}
\title{create correlation plot}
\usage{
create_correlation_plot(matrix_data, groups = NULL,
  min_correlation = 0.5)
}
\arguments{
\item{matrix_data}{the matrix of data, variables by samples (r x c)}

\item{groups}{the groups of the samples}

\item{min_correlation}{the minimum correlation value to color}
}
\description{
Given matrix data, create a correlation heatmap with reordering by group.
}
