% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{create_value}
\alias{create_value}
\title{create a value from mantissa and exponent}
\usage{
create_value(mantissa, exponent)
}
\arguments{
\item{mantissa}{the base part of the number}

\item{exponent}{the exponent part}
}
\value{
numeric
}
\description{
given a mantissa and exponent, returns the actual value as a numeric
}

