% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{extract}
\alias{extract}
\title{extract parts of a value}
\usage{
extract(x)
}
\arguments{
\item{x}{the number to extract the parts from}
}
\value{
list
}
\description{
Often we want to transform a number into it's exponential representation,
having the number itself and the number of decimal places. This function
provides that functionality
}

