% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{log_with_min}
\alias{log_with_min}
\title{log-transform data}
\usage{
log_with_min(data, min_value = NULL, order_mag = 3, log_fun = log)
}
\arguments{
\item{data}{the data to work with}

\item{min_value}{the minimum value}

\item{order_mag}{how many orders of magnitute smaller should min value be?}

\item{log_fun}{what log function to use for the transformation}
}
\value{
matrix
}
\description{
performs a log-transform while adding a small value to the data based on
finding the smallest non-zero value in the data
}

