% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid-up-down-plots.R
\name{mu_add_annotations}
\alias{mu_add_annotations}
\title{add annotations}
\usage{
mu_add_annotations(log_df, feature_annotations)
}
\arguments{
\item{log_df}{data.frame of log-fold-changes}

\item{feature_annotations}{a categoryCompare feature annotations object}
}
\value{
list
}
\description{
Given a data.frame with `feature_id` and `log2FoldChange`, and an annotation
object with lipid classes, `total_length`, `total_db`, `chain_length`, and `chain_db`,
constructs a data.frame of annotations, and prepares the log-fold-change for plotting.
}
