% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{mu_apply_normalization}
\alias{mu_apply_normalization}
\title{apply normalization}
\usage{
mu_apply_normalization(intensity_matrix, normalization_factors,
  transform = NULL)
}
\arguments{
\item{intensity_matrix}{matrix of intensities, rows are entries, columns are samples}

\item{normalization_factors}{named vector of normalization factors}
}
\value{
matrix
}
\description{
Apply normalization to a matrix of intensities. This uses division, so
if you had something logged, you would want to do the exponent first.
}
