% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{mu_cooccurence_distance}
\alias{mu_cooccurence_distance}
\title{get cooccurence distance}
\usage{
mu_cooccurence_distance(intensity_matrix, threshold_value = 0, method = 2)
}
\arguments{
\item{intensity_matrix}{matrix of intensity or propensity values}

\item{threshold_value}{the value for thresholding to zero}

\item{method}{the co-ocurrence method to use (see ade4::dist.binary), default is "2"}
}
\value{
matrix
}
\description{
Calculates a binary distance metric
}
