% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{mu_cooccurence_heatmap}
\alias{mu_cooccurence_heatmap}
\title{create cooccurrence heatmap}
\usage{
mu_cooccurence_heatmap(distance_matrix, groups = NULL, min_value = 0)
}
\arguments{
\item{distance_matrix}{matrix of binary distances}

\item{groups}{data.frame of groups (default = NULL)}

\item{min_value}{the minimum value to use (default = 0)}
}
\description{
given a distance generated by `mu_cooccurrence_distance`, generate the
corresponding heatmap.
}
