% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf.R
\name{mu_emf_sum}
\alias{mu_emf_sum}
\title{calculate emf sums}
\usage{
mu_emf_sum(normalized_data, emf, peak_info = NULL)
}
\arguments{
\item{normalized_data}{a matrix of normalized intensities}

\item{emf}{which emf are we working with}

\item{peak_info}{an optional data.frame of peak information}
}
\value{
A data.frame or list
}
\description{
For a given set of EMF data, calculates a nominal sum, or a set of weighted
sums where the weights are based on the number of C13 incorporations.
}
\details{
If `peak_info` is `NULL` (not supplied), then the *nominal* sum is
  returned. If `peak_info` is supplied, then the weighted sums for each different
  EMF is returned. Note that the rownames of `normalized_data` should correspond
  to the `PeakID` in `peak_info`, and `PeakID`, `c13`, `complete_EMF` and `complete_IMF` must
  all be present.
}
