% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf.R
\name{mu_normalize_emf}
\alias{mu_normalize_emf}
\title{normalize emf}
\usage{
mu_normalize_emf(emf, normalization_factors, threshold_value = NULL)
}
\arguments{
\item{emf}{the EMF matrix of values}

\item{normalization_factors}{the vector of normalization values}

\item{threshold_value}{the threshold value to replace NA's}
}
\value{
matrix
}
\description{
Given an EMF matrix, normalize and replace NAs with a threshold value
}
