% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf.R
\name{mu_numberc13}
\alias{mu_numberc13}
\title{get number of c13 incorporations}
\usage{
mu_numberc13(imfs, add_one = FALSE)
}
\arguments{
\item{imfs}{vector of character formulas}

\item{add_one}{should **1** be added (to avoid zero)?}
}
\value{
integer
}
\description{
Given a vector of isotopic molecular formula's (IMFs), attempts to determine
how many C13 incorporations are present. Isotopic element counts should be
separated by ","
}
