% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{mu_peak_intensity}
\alias{mu_peak_intensity}
\title{return all peak heights}
\usage{
mu_peak_intensity(zip_file)
}
\arguments{
\item{zip_file}{the zip file for a sample}
}
\value{
data.frame
}
\description{
return all peak heights
}
