% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid-up-down-plots.R
\name{mu_plot_up_down_classes}
\alias{mu_plot_up_down_classes}
\title{create an up-down plot of classes}
\usage{
mu_plot_up_down_classes(class_summary, class_order = NULL)
}
\arguments{
\item{class_summary}{the summary from `summarize_up_down`}

\item{class_order}{the subset / ordering of classes}
}
\value{
ggplot2
}
\description{
Generate an up-down changed barplot for classes.
}
