% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid-up-down-plots.R
\name{mu_plot_up_down_length_db}
\alias{mu_plot_up_down_length_db}
\title{up-down plot within a class}
\usage{
mu_plot_up_down_length_db(other_summary, which_class = NULL)
}
\arguments{
\item{other_summary}{the summary from `summarize_up_down`}

\item{which_class}{which class to create the plots for}
}
\value{
ggplot2
}
\description{
Generate sets of up-down plots within classes
}
