% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf.R
\name{mu_ratio_emfSum}
\alias{mu_ratio_emfSum}
\title{ratio sum}
\usage{
mu_ratio_emfSum(emf)
}
\arguments{
\item{emf}{the EMF matrix of values}
}
\value{
matrix
}
\description{
Given an EMF matrix, take the ratio of all peaks to the sum of the peaks
}
