% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf.R
\name{mu_ratio_m0}
\alias{mu_ratio_m0}
\title{ratio m0}
\usage{
mu_ratio_m0(emf)
}
\arguments{
\item{emf}{the EMF matrix of values}
}
\value{
matrix
}
\description{
Given an EMF matrix, take the ratio of all m+ peaks to the m0 peak to compare c13 incorporation
}
