% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_mz_ranges.R
\name{mz_to_iranges}
\alias{mz_to_iranges}
\title{mz ranges to IRanges}
\usage{
mz_to_iranges(mz_values, multi_factor = 10000)
}
\arguments{
\item{mz_values}{either a vector or matrix of single or double ranges}

\item{multi_factor}{how much to multiply values by}
}
\value{
IRanges
}
\description{
modifies a matrix of m/z ranges into an IRanges object to calculate
overlaps. Because IRanges objects are integer based, also multiplies
by a factor (default 10,000) to preserve resolution.
}

