% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{ttest_matrix}
\alias{ttest_matrix}
\title{matrix t-test}
\usage{
ttest_matrix(data_matrix, class_data, adjust_method = "BH", ...)
}
\arguments{
\item{data_matrix}{the matrix of data we are working with}

\item{class_data}{the classes of the columns}

\item{adjust_method}{how to adjust the p.values}

\item{...}{other parameters for t.test}
}
\value{
data.frame
}
\description{
matrix t-test
}
