% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{ttest_vector}
\alias{ttest_vector}
\title{t-test}
\usage{
ttest_vector(data_vector, class_data, ...)
}
\arguments{
\item{data_vector}{the vector we are working with}

\item{class_data}{what class are each entry in}

\item{...}{other parameters for t.test}
}
\value{
data.frame
}
\description{
runs t-tests on a vector, with classes of the vector supplied.
}
