% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_voting.R
\name{vote_categories_classes}
\alias{vote_categories_classes}
\title{vote on categories and classes}
\usage{
vote_categories_classes(
  categories_classes,
  emfs = "emf",
  categories = "Categories",
  classes = "Classes",
  return_opts = "max",
  vote_threshold = 2/3 * 100,
  exclude_votes = c("not_lipid", "unclassifiable", as.character(NA),
    "not_lipid or not_classified"),
  debug = FALSE
)
}
\arguments{
\item{categories_classes}{data.frame of information}

\item{emfs}{which column is EMFs}

\item{categories}{which column is categories}

\item{classes}{which column has classes}

\item{return_opts}{what to return (`max`, `all`, `single`, `multiple`)}

\item{vote_threshold}{what is the vote threshold for `single`?}

\item{exclude_votes}{which categories should be excluded?}
}
\value{
data.frame
}
\description{
Given a data.frame of categories, classes, and EMFs, does voting by the EMF first
on the Categories, and then trims to those Classes that are part of the original Category,
and does voting on the remaining Classes.
}
