% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xlsx_to_json}
\alias{xlsx_to_json}
\title{convert peaks}
\usage{
xlsx_to_json(in_xlsx, save_loc)
}
\arguments{
\item{in_xlsx}{the file to work on}

\item{save_loc}{the directory to save the json file}
}
\description{
converts an XLSX peak list from Xcalibur to a json peak format.
}

