% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_classifications.R
\name{import_emf_classifications}
\alias{import_emf_classifications}
\title{Import EMF Classifications}
\usage{
import_emf_classifications(classification_file, fix_json = FALSE)
}
\arguments{
\item{classification_file}{the file to read from}

\item{fix_json}{properly format the JSON so the JSON parser can read it?}

\item{rm_na_formula}{remove those entries that have a formula entry of `NA`??}
}
\value{
data.frame
}
\description{
Allows the reading in of EMF classification results from a JSON file.
}
