% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{mu_heatmap}
\alias{mu_heatmap}
\title{create a heatmap}
\usage{
mu_heatmap(matrix_data, title = "", groups = NULL,
  transform = "none", min_value = 0)
}
\arguments{
\item{matrix_data}{matrix of binary distances}

\item{title}{the title of the data}

\item{groups}{data.frame of groups (default = NULL)}

\item{transform}{how to transform the data for similarity reordering}

\item{min_value}{the minimum value to use (default = 0)}
}
\description{
given a matrix denoting either distance or similarity, create a heatmap.
}
